# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "DIMO Devices API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
        "/device-definitions": {
            "get": {
                "description": "gets a specific device definition by make model and year",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-definitions"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "make eg TESLA",
                        "name": "make",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "model eg MODEL Y",
                        "name": "model",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "year eg 2021",
                        "name": "year",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.DeviceDefinition"
                        }
                    }
                }
            }
        },
...
        "/device-definitions/{id}": {
            "get": {
                "description": "gets a specific device definition by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-definitions"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "device definition id, KSUID format",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.DeviceDefinition"
                        }
                    }
                }
            }
        },
        "/device-definitions/{id}/integrations": {
            "get": {
                "description": "gets all the available integrations for a device definition. Includes the capabilities of the device with the integration",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-definitions"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "device definition id, KSUID format",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/services.DeviceCompatibility"
                            }
                        }
                    }
                }
            }
        },
...
        "/user/device-data/{userDeviceID}/distance-driven": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get kilometers driven for a userDeviceID since connected (ie. since we have data available)\nif it returns 0 for distanceDriven it means we have no odometer data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user device id",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "404": {
                        "description": "no device found for user with provided parameters"
                    }
                }
            }
        },
        "/user/device-data/{userDeviceID}/historical": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get historical data for a userDeviceID, within start and end range, taking the\nlatest status from every 30m bucket.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "startDate eg 2022-01-02. if empty two weeks back",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "endDate eg 2022-03-01. if empty today",
                        "name": "endDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "odometer. if empty all",
                        "name": "signal",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "intervals for bucketing, 30m, 60m. if empty 30m",
                        "name": "interval",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "invalid start or end date"
                    },
                    "404": {
                        "description": "no device found for user with provided parameters"
                    }
                }
            }
        },
        "/user/device-data/{userDeviceID}/historical-30m": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get historical data for a userDeviceID, within start and end range, taking the\nlatest status from every 30m bucket.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "device-data"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "startDate eg 2022-01-02. if empty two weeks back",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "endDate eg 2022-03-01. if empty today",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "invalid start or end date"
                    },
                    "404": {
                        "description": "no device found for user with provided parameters"
                    }
                }
            }
        },
        "/user/devices": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "adds a device to a user. can add with only device_definition_id or with MMY, which will create a device_definition on the fly",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-devices"
                ],
                "parameters": [
                    {
                        "description": "add device to user. either MMY or id are required",
                        "name": "user_device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github.com_DIMO-Network_devices-api_internal_controllers.RegisterUserDevice"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github.com_DIMO-Network_devices-api_internal_controllers.RegisterUserDeviceResponse"
                        }
                    }
                }
            }
        },
        "/user/devices/:userDeviceID/autopi/command": {
            "post": {
                "description": "Submit a raw autopi command to unit. Device must be registered with autopi before this can be used",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "parameters": [
                    {
                        "description": "raw autopi command",
                        "name": "AutoPiCommandRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.AutoPiCommandRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/user/devices/:userDeviceID/autopi/command/:jobID": {
            "get": {
                "description": "gets the status of an autopi raw command by jobID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "job id, from autopi",
                        "name": "jobID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.UserDeviceAPIIntegrationJob"
                        }
                    }
                }
            }
        },
        "/user/devices/:userDeviceID/integrations/:integrationID": {
            "post": {
                "description": "Submit credentials for registering a device with a given integration.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "integrations"
                ],
                "parameters": [
                    {
                        "description": "Integration credentials",
                        "name": "userDeviceIntegrationRegistration",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers.RegisterDeviceIntegrationRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
        "/user/devices/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "gets all devices associated with current user - pulled from token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user-devices"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github.com_DIMO-Network_devices-api_internal_controllers.UserDeviceFull"
                            }
                        }
                    }
                }
            }
        },
        "/user/devices/{userDeviceID}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "delete the user device record (hard delete)",
                "tags": [
                    "user-devices"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "userDeviceID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
        "/user/devices/{userDeviceID}/commands/refresh": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Starts the process of refreshing device status from Smartcar",
                "tags": [
                    "user-devices"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "user device ID",
                        "name": "user_device_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "429": {
                        "description": "rate limit hit for integration"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_controllers.AutoPiCommandRequest": {
            "type": "object",
            "properties": {
                "command": {
                    "type": "string"
                }
            }
        },
...
        "services.DeviceCompatibility": {
            "type": "object",
            "properties": {
                "capabilities": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "style": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "vendor": {
                    "type": "string"
                }
            }
        },
        "services.DeviceDefinition": {
            "type": "object",
            "properties": {
                "compatibleIntegrations": {
                    "description": "CompatibleIntegrations has systems this vehicle can integrate with",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.DeviceCompatibility"
                    }
                },
                "deviceDefinitionId": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "metadata": {},
                "name": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/services.DeviceType"
                },
                "vehicleData": {
                    "description": "VehicleInfo will be empty if not a vehicle type",
                    "$ref": "#/definitions/services.DeviceVehicleInfo"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}